// src/store/main/system/types.ts
export interface ISystemState {
  usersList: any[]
  usersCount: number
  departmentList: any[]
  departmentCount: number
  roleList: any[]
  roleCount: number
  goodsList: any[]
  goodsCount: number
  categoryList: any[]
  categoryCount: number
  menuList: any[]
  menuCount: number
  storyList: any[]
  storyCount: number
}